Move event slicing to before adjustment workspaces
This produces same output as old algorithm
