+++ Framework/PythonInterface/plugins/algorithms/WorkflowAlgorithms/SANS/SANSReductionCoreEventSlice.py
@@ -17,8 +17,7 @@ from mantid.kernel import (Direction, PropertyManagerProperty, StringListValidat
 from sans.algorithm_detail.mask_workspace import mask_bins
 from sans.common.constants import EMPTY_NAME
 from sans.common.enums import (DetectorType, DataType)
-from sans.common.general_functions import (append_to_sans_file_tag, create_child_algorithm,
-                                           create_managed_non_child_algorithm)
+from sans.common.general_functions import append_to_sans_file_tag, create_child_algorithm
 from sans.state.state_base import create_deserialized_sans_state_from_property_manager
 
 
@@ -49,24 +48,20 @@ class SANSReductionCoreEventSlice(DistributedDataProcessorAlgorithm):
         self.declareProperty(MatrixWorkspaceProperty('ScatterMonitorWorkspace', '',
                                                      optional=PropertyMode.Optional, direction=Direction.Input),
                              doc='The scatter monitor workspace. This workspace only contains monitors.')
-        # Wavelength Adjustment
-        self.declareProperty(MatrixWorkspaceProperty('WavelengthAdjustmentWorkspace', '',
+        # Direct Workspace
+        self.declareProperty(MatrixWorkspaceProperty('DirectWorkspace', '',
                                                      optional=PropertyMode.Optional, direction=Direction.Input),
                              doc='The direct workspace.')
-        # Pixel Adjustment
-        self.declareProperty(MatrixWorkspaceProperty('PixelAdjustmentWorkspace', '',
+        # Transmission Workspace
+        self.declareProperty(MatrixWorkspaceProperty('TransmissionWorkspace', '',
                                                      optional=PropertyMode.Optional, direction=Direction.Input),
-                             doc='The direct workspace.')
-        # Wavelength and Pixel Adjustment
-        self.declareProperty(MatrixWorkspaceProperty('WavelengthAndPixelAdjustmentWorkspace', '',
-                                                     optional=PropertyMode.Optional, direction=Direction.Input),
-                             doc='The direct workspace.')
+                             doc='The transmission workspace')
 
         self.setPropertyGroup("ScatterWorkspace", 'Data')
         self.setPropertyGroup("ScatterMonitorWorkspace", 'Data')
-        self.setPropertyGroup("WavelengthAdjustmentWorkspace", 'Data')
-        self.setPropertyGroup("PixelAdjustmentWorkspace", 'Data')
-        self.setPropertyGroup("WavelengthAndPixelAdjustmentWorkspace", 'Data')
+        self.setPropertyGroup("DummyMaskWorkspace", 'Data')
+        self.setPropertyGroup("DirectWorkspace", 'Data')
+        self.setPropertyGroup("TransmissionWorkspace", 'Data')
 
         # The component
         allowed_detectors = StringListValidator([DetectorType.to_string(DetectorType.LAB),
@@ -96,6 +91,14 @@ class SANSReductionCoreEventSlice(DistributedDataProcessorAlgorithm):
                                                      direction=Direction.Output),
                              doc='The sum of the counts of the output workspace.')
 
+        self.declareProperty(MatrixWorkspaceProperty('CalculatedTransmissionWorkspace', '', optional=PropertyMode.Optional,
+                                                     direction=Direction.Output),
+                             doc='The calculated transmission workspace')
+
+        self.declareProperty(MatrixWorkspaceProperty('UnfittedTransmissionWorkspace', '', optional=PropertyMode.Optional,
+                                                     direction=Direction.Output),
+                             doc='The unfitted transmission workspace')
+
     def PyExec(self):
         # Get the input
         state = self._get_state()
@@ -114,6 +117,23 @@ class SANSReductionCoreEventSlice(DistributedDataProcessorAlgorithm):
         workspace, monitor_workspace, slice_event_factor = self._slice(state_serialized, workspace, monitor_workspace,
                                                                        data_type_as_string)
 
+        # --------------------------------------------------------------------------------------------------------------
+        # 7. Create adjustment workspaces, those are
+        #     1. pixel-based adjustments
+        #     2. wavelength-based adjustments
+        #     3. pixel-and-wavelength-based adjustments
+        # Note that steps 4 to 7 could run in parallel if we don't use wide angle correction. If we do then the
+        # creation of the adjustment workspaces requires the sample workspace itself and we have to run it sequentially.
+        # We could consider to have a serial and a parallel strategy here, depending on the wide angle correction
+        # settings. On the other hand it is not clear that this would be an advantage with the GIL.
+        # --------------------------------------------------------------------------------------------------------------
+        component_as_string = self.getProperty("Component").value
+        data_type_as_string = self.getProperty("DataType").value
+        progress.report("Creating adjustment workspaces ...")
+        wavelength_adjustment_workspace, pixel_adjustment_workspace, wavelength_and_pixel_adjustment_workspace, \
+            calculated_transmission_workspace, unfitted_transmission_workspace = \
+            self._adjustment(state_serialized, workspace, monitor_workspace, component_as_string, data_type_as_string)
+
         # ------------------------------------------------------------
         # 2. Convert event workspaces to histogram workspaces
         # ------------------------------------------------------------
@@ -121,25 +141,22 @@ class SANSReductionCoreEventSlice(DistributedDataProcessorAlgorithm):
         workspace = self._convert_to_histogram(workspace)
 
         # ------------------------------------------------------------
+        # 3. Re-mask. We need to bin mask in histogram mode in order
+        #    to have knowledge of masked regions: masking
+        #    EventWorkspaces simply removes their events
+        # ------------------------------------------------------------
+        dummy_mask_workspace = self.getProperty("DummyMaskWorkspace").value
+        workspace = self._copy_bin_masks(workspace, dummy_mask_workspace)
+
+        # ------------------------------------------------------------
         # 3. Convert to Q
         # -----------------------------------------------------------
         progress.report("Converting to q ...")
-        wavelength_adjustment_workspace = self.getProperty("WavelengthAdjustmentWorkspace").value
-        pixel_adjustment_workspace = self.getProperty("PixelAdjustmentWorkspace").value
-        wavelength_and_pixel_adjustment_workspace = self.getProperty("WavelengthAndPixelAdjustmentWorkspace").value
         workspace, sum_of_counts, sum_of_norms = self._convert_to_q(state_serialized,
                                                                     workspace,
                                                                     wavelength_adjustment_workspace,
                                                                     pixel_adjustment_workspace,
                                                                     wavelength_and_pixel_adjustment_workspace)
-
-        # ------------------------------------------------------------
-        # 4. Re-mask. We need to bin mask in histogram mode in order
-        #    to have knowledge of masked regions: masking
-        #    EventWorkspaces simply removes their events
-        # ------------------------------------------------------------
-        dummy_mask_workspace = self.getProperty("DummyMaskWorkspace").value
-        self._copy_bin_masks(workspace, dummy_mask_workspace)
 
         progress.report("Completed SANSReductionCoreEventSlice...")
 
@@ -156,6 +173,9 @@ class SANSReductionCoreEventSlice(DistributedDataProcessorAlgorithm):
         if sum_of_norms:
             self.setProperty("SumOfNormFactors", sum_of_norms)
 
+        self.setProperty("CalculatedTransmissionWorkspace", calculated_transmission_workspace)
+        self.setProperty("UnfittedTransmissionWorkspace", unfitted_transmission_workspace)
+
     def _slice(self, state_serialized, workspace, monitor_workspace, data_type_as_string):
         slice_name = "SANSSliceEvent"
         slice_options = {"SANSState": state_serialized,
@@ -171,6 +191,68 @@ class SANSReductionCoreEventSlice(DistributedDataProcessorAlgorithm):
         monitor_workspace = slice_alg.getProperty("OutputWorkspaceMonitor").value
         slice_event_factor = slice_alg.getProperty("SliceEventFactor").value
         return workspace, monitor_workspace, slice_event_factor
+
+    def _adjustment(self, state_serialized, workspace, monitor_workspace, component_as_string, data_type):
+        transmission_workspace = self._get_transmission_workspace()
+        direct_workspace = self._get_direct_workspace()
+
+        adjustment_name = "SANSCreateAdjustmentWorkspaces"
+        adjustment_options = {"SANSState": state_serialized,
+                              "Component": component_as_string,
+                              "DataType": data_type,
+                              "MonitorWorkspace": monitor_workspace,
+                              "SampleData": workspace,
+                              "OutputWorkspaceWavelengthAdjustment": EMPTY_NAME,
+                              "OutputWorkspacePixelAdjustment": EMPTY_NAME,
+                              "OutputWorkspaceWavelengthAndPixelAdjustment": EMPTY_NAME}
+        if transmission_workspace:
+            transmission_workspace = self._move(state_serialized, transmission_workspace, component_as_string,
+                                                is_transmission=True)
+            adjustment_options.update({"TransmissionWorkspace": transmission_workspace})
+
+        if direct_workspace:
+            direct_workspace = self._move(state_serialized, direct_workspace, component_as_string, is_transmission=True)
+            adjustment_options.update({"DirectWorkspace": direct_workspace})
+
+        adjustment_alg = create_child_algorithm(self, adjustment_name, **adjustment_options)
+        adjustment_alg.execute()
+
+        wavelength_adjustment = adjustment_alg.getProperty("OutputWorkspaceWavelengthAdjustment").value
+        pixel_adjustment = adjustment_alg.getProperty("OutputWorkspacePixelAdjustment").value
+        wavelength_and_pixel_adjustment = adjustment_alg.getProperty(
+                                           "OutputWorkspaceWavelengthAndPixelAdjustment").value
+        calculated_transmission_workspace = adjustment_alg.getProperty("CalculatedTransmissionWorkspace").value
+        unfitted_transmission_workspace = adjustment_alg.getProperty("UnfittedTransmissionWorkspace").value
+        return wavelength_adjustment, pixel_adjustment, wavelength_and_pixel_adjustment, \
+            calculated_transmission_workspace, unfitted_transmission_workspace
+
+    def _move(self, state_serialized, workspace, component, is_transmission=False):
+        # First we set the workspace to zero, since it might have been moved around by the user in the ADS
+        # Second we use the initial move to bring the workspace into the correct position
+        move_name = "SANSMove"
+        move_options = {"SANSState": state_serialized,
+                        "Workspace": workspace,
+                        "MoveType": "SetToZero",
+                        "Component": ""}
+        move_alg = create_child_algorithm(self, move_name, **move_options)
+        move_alg.execute()
+        workspace = move_alg.getProperty("Workspace").value
+
+        # Do the initial move
+        move_alg.setProperty("MoveType", "InitialMove")
+        move_alg.setProperty("Component", component)
+        move_alg.setProperty("Workspace", workspace)
+        move_alg.setProperty("IsTransmissionWorkspace", is_transmission)
+        move_alg.execute()
+        return move_alg.getProperty("Workspace").value
+
+    def _get_transmission_workspace(self):
+        transmission_workspace = self.getProperty("TransmissionWorkspace").value
+        return self._get_cloned_workspace(transmission_workspace) if transmission_workspace else None
+
+    def _get_direct_workspace(self):
+        direct_workspace = self.getProperty("DirectWorkspace").value
+        return self._get_cloned_workspace(direct_workspace) if direct_workspace else None
 
     def _convert_to_histogram(self, workspace):
         if isinstance(workspace, IEventWorkspace):
